feat: refactor useColorScheme hook to improve color scheme management; consolidate imports and enhance persistence
