+++ src/components/ui/activity-indicator.tsx
@@ -1,7 +1,6 @@
+import { useColorScheme } from "@/src/hooks/useColorScheme";
 import { ActivityIndicator as RNActivityIndicator } from "react-native";
 import { NAV_THEME } from "~/lib/constants";
-
-import { useColorScheme } from "~/lib/useColorScheme";
 
 function ActivityIndicator(
 	props: React.ComponentPropsWithoutRef<typeof RNActivityIndicator>,
